==> 4_Quora_FuzzywuzzyFeat_Train.ipynb <==
import numpy as np
import pandas as pd
import sklearn
%matplotlib inline
import matplotlib.pyplot as plt
from fuzzywuzzy import fuzz

## Train Data Features

# Loadin Train Data with basic features
train_data1 = pd.read_csv('train_data1.csv', sep=',')
train_data2 = pd.read_csv('train_data2.csv', sep=',', skiprows=1, header=None)

# Creating DF with features and duplicate variable only for final model
def append_df(df1, df2):
    # df1 with column names, and df2 without
    df2.columns = df1.columns
    df1 = df1.append(df2, ignore_index=True)
    return df1

train_data1 = append_df(train_data1,train_data2)
train_data1.shape

train_data1.count()

train_data1['prcnt_common'].fillna(train_data1['prcnt_common'].mean(), inplace=True)

train_data1['question1_form'].fillna('', inplace=True)
train_data1['question2_form'].fillna('', inplace=True)

train_data1.count()

#Empty cols in the training data
train_data1[(train_data1['question2'].isnull())]

train_data1.question2[(train_data1['question2'].isnull())]=''

def fuzzy_ratios(input_df):
    input_df['partial_ratio'] = input_df.apply(lambda row: fuzz.partial_ratio(row[u'question1_form'], row[u'question2_form']), axis=1)
    input_df['partial_token_set_ratio'] = input_df.apply(lambda row: fuzz.partial_token_set_ratio(row[u'question1_form'], row[u'question2_form']), axis=1)
    input_df['partial_token_sort_ratio'] = input_df.apply(lambda row: fuzz.partial_token_sort_ratio(row[u'question1_form'], row[u'question2_form']), axis=1)
    input_df['q_ratio'] = input_df.apply(lambda row: fuzz.QRatio(row[u'question1_form'], row[u'question2_form']), axis=1)
    input_df['token_set_ratio'] = input_df.apply(lambda row: fuzz.token_set_ratio(row[u'question1_form'], row[u'question2_form']), axis=1)
    input_df['token_sort_ratio'] = input_df.apply(lambda row: fuzz.token_sort_ratio(row[u'question1_form'], row[u'question2_form']), axis=1)
    input_df['w_ratio'] = input_df.apply(lambda row: fuzz.WRatio(row[u'question1_form'], row[u'question2_form']), axis=1)
#   Issue of the Non-ASCII Characters with the below two ratios, ignored as of now. Col_clean function for it, if to be used    
#    input_df['uw_ratio'] = input_df.apply(lambda row: fuzz.UWRatio(row[u'question1_form'], row[u'question2_form']), axis=1)
#    input_df['uq_ratio'] = input_df.apply(lambda row: fuzz.UQRatio(row[u'question1_form'], row[u'question2_form']), axis=1)
    return input_df

import re
# Removing the non-ascii characters
def col_clean(input_col):
    input_col['question1_form'] = input_col.apply(lambda row: re.sub( r'[^\x00-\x7F]+','', row[u'question1_form']), axis=1)
    input_col['question1_form'] = input_col.apply(lambda row: re.sub(r'(?:\s{2,})',' ', row[u'question1_form']), axis=1)
    input_col['question1_form'] = input_col.apply(lambda row: row[u'question1_form'].rstrip(), axis=1)
    input_col['question1_form'] = input_col.apply(lambda row: row[u'question1_form'].lstrip(), axis=1)
    return input_col

# Not used as of now!!
#train_data1[['question1_form']]=col_clean(train_data1[['question1_form']])
#train_data1[['question2_form']]=col_clean(train_data1[['question2_form']])

train_data1.count()

# Storing the file as csv for the XGB Model
train_data_save = train_data1
train_data_save.drop(['id','qid1','qid2','question1','question2'], axis=1, inplace=True)
train_data_save.to_csv('train_basic_fuzzy.csv',index=False)
